(feat) move around files and make it work
