+++ hummingbot/cli/hummingbot_application.py
@@ -47,15 +47,20 @@ from hummingbot.cli.errors import (
     ArgumentParserError
 )
 from hummingbot.cli.config.config_var import ConfigVar
-from hummingbot.cli.settings import (
+from hummingbot.cli.config.in_memory_config_map import (
     in_memory_config_map,
-    global_config_map,
     get_strategy_config_map,
     write_config_to_yml,
     load_required_configs,
-    EXCHANGES,
+)
+from hummingbot.cli.config.global_config_map import global_config_map
+from hummingbot.cli.config.helpers import (
     parse_cvar_value,
     copy_strategy_template,
+    get_erc20_token_addresses,
+)
+from hummingbot.cli.settings import (
+    EXCHANGES,
 )
 from hummingbot.logger.report_aggregator import ReportAggregator
 from hummingbot.strategy.cross_exchange_market_making import (
@@ -66,7 +71,6 @@ from hummingbot.strategy.arbitrage import (
     ArbitrageStrategy,
     ArbitrageMarketPair
 )
-from hummingbot.cli.settings import get_erc20_token_addresses
 from hummingbot.cli.utils.exchange_rate_conversion import ExchangeRateConversion
 
 s_logger = None
