+++ hummingbot/logger/reporting_proxy_handler.py
@@ -1,7 +1,7 @@
 from typing import Optional
 import logging
 import json
-from hummingbot.cli.settings import global_config_map
+from hummingbot.cli.config.global_config_map import global_config_map
 from wings.logger.struct_logger import log_encoder
 from hummingbot.logger.log_server_client import LogServerClient
 from hummingbot.logger.report_aggregator import REPORT_EVENT_QUEUE
